+++ doc/changes.rst
@@ -2,10 +2,44 @@
 desisim change log
 ==================
 
-0.31.1 (unreleased)
+0.31.3 (unreleased)
 -------------------
 
-* No changes yet
+* No changes yet.
+
+0.31.2 (2019-02-28)
+-------------------
+
+* Update quickquasars default redshift error (`PR #466`_).
+* Support for London mocks v5.0 + DLAs (`PR #467`_).
+
+.. _`PR #466`: https://github.com/desihub/desisim/pull/466
+.. _`PR #467`: https://github.com/desihub/desisim/pull/467
+
+0.31.1 (2018-12-14)
+-------------------
+
+* quickquasars updates:
+
+  * support eBOSS (`PR #450`_).
+  * mimic redshift fitter uncertainties (`PR #452`_).
+  * adding shift to redshift (`PR #454`_).
+  * fix error in size of Z_noFOG (`PR #455`_).
+  * Fix quickquasars targetid truth (`PR #457`_).
+
+* Precompute colors for star and galaxy templates. (`PR #453`_).
+* Refactor S/N qa to load cframes only once (also updates OII for new TRUTH table) (`PR #459`_, `PR #465`_).
+* Use basis_templates v3.1 and matching desisim-testdata 0.6.1 (`PR #464`_).
+
+.. _`PR #450`: https://github.com/desihub/desisim/pull/450
+.. _`PR #452`: https://github.com/desihub/desisim/pull/452
+.. _`PR #453`: https://github.com/desihub/desisim/pull/453
+.. _`PR #454`: https://github.com/desihub/desisim/pull/454
+.. _`PR #455`: https://github.com/desihub/desisim/pull/455
+.. _`PR #457`: https://github.com/desihub/desisim/pull/457
+.. _`PR #459`: https://github.com/desihub/desisim/pull/459
+.. _`PR #464`: https://github.com/desihub/desisim/pull/464
+.. _`PR #465`: https://github.com/desihub/desisim/pull/465
 
 0.31.0 (2018-11-08)
 -------------------